+++ pkg/jwt/logger.go
@@ -5,4 +5,5 @@ import "context"
 type logger interface {
 	ErrUserAuth(context.Context, error) error
 	InfoGeneric(context.Context, string) error
+	ErrGeneric(context.Context, error) error
 }