+++ pkg/jwt/middleware/middleware.go
@@ -13,6 +13,7 @@ import (
 type logger interface {
 	ErrUserAuth(context.Context, error) error
 	InfoGeneric(context.Context, string) error
+	ErrGeneric(context.Context, error) error
 }
 
 // Middleware is the default middleware type used everywhere.