+++ pkg/jwt/test/logger.go
@@ -15,3 +15,7 @@ func (Logger) InfoGeneric(ctx context.Context, msg string) error {
 	fmt.Println(msg)
 	return nil
 }
+func (Logger) ErrGeneric(ctx context.Context, err error) error {
+	fmt.Println(err)
+	return nil
+}