+++ pkg/log/http.go
@@ -9,6 +9,8 @@ import (
 	"regexp"
 	"strings"
 	"time"
+
+	"github.com/gesundheitscloud/go-svc/pkg/d4lcontext"
 )
 
 type contextKey string
@@ -83,9 +85,13 @@ func (l HTTPLogger) ServeHTTP(rw http.ResponseWriter, req *http.Request) {
 
 	if userID := l.userParser(req); userID != "" {
 		req = req.WithContext(context.WithValue(req.Context(), UserIDContextKey, userID))
+	} else if userID = d4lcontext.GetUserID(req); userID != "" {
+		req = req.WithContext(context.WithValue(req.Context(), UserIDContextKey, userID))
 	}
 
 	if clientID := l.clientParser(req); clientID != "" {
+		req = req.WithContext(context.WithValue(req.Context(), ClientIDContextKey, clientID))
+	} else if clientID = d4lcontext.GetClientID(req); clientID != "" {
 		req = req.WithContext(context.WithValue(req.Context(), ClientIDContextKey, clientID))
 	}
 
@@ -94,6 +100,8 @@ func (l HTTPLogger) ServeHTTP(rw http.ResponseWriter, req *http.Request) {
 	}
 
 	if tenantID := l.tenantIDParser(req); tenantID != "" {
+		req = req.WithContext(context.WithValue(req.Context(), TenantIDContextKey, tenantID))
+	} else if tenantID = d4lcontext.GetTenantID(req); tenantID != "" {
 		req = req.WithContext(context.WithValue(req.Context(), TenantIDContextKey, tenantID))
 	}