+++ CHANGELOG.md
@@ -9,13 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- JWT pkg - `Extract` middleware that extracts the JWT information without doing any access control
 - JWT pkg - function for creating a signed access token
 - JWT pkg - token for mail verification
+- d4lcontext pkg - methods allowing to add values to a request's context using the d4lcontext keys
 - DB pkg - use migrate pkg after the execution of the gorm migrations
 
 ### Changed
 
 - [breaking] Move `ParseRequesterID` function from `d4lcontext` to `d4lhandler` package
+- Log pkg - support the `d4lcontext` keys as a fallback for user ID, client ID and tenant ID
+- d4lcontext pkg - GetTenantID doesn't fall back to 'd4l' if the tenant ID is missing in the context
 
 ### Deprecated